+++ src/objects/QueueManager.ts
